refec : categories -> aritcles 하위에서 독립 앱으로 분화
