+++ articles/serializers.py
@@ -1,7 +1,8 @@
 from rest_framework.serializers import ModelSerializer
 
 from users.serializers import UserNameSerializer
-from .models import Article, Category
+from .models import Article
+from categories.models import Category
 
 class CategoryNameSerializer(ModelSerializer):
     class Meta:
@@ -28,19 +29,4 @@ class ArticleSerializer(ModelSerializer):
         extra_kwargs = {
             'author' : {"read_only" : True},
             'category' : {"read_only" : True}
-        }
-        
-class CategorySerializer(ModelSerializer):
-    
-    class Meta:
-        model = Category
-        fields = (
-            'category_id',
-            'category_name',
-            'is_public',
-            'created_user',
-            'parent',
-        )
-        extra_kwargs = {
-            "created_user" : {"read_only" : True}
         }
