+++ config/settings/base.py
@@ -43,6 +43,7 @@ INSTALLED_APPS = [
 
     'users',
     'articles',
+    'categories',
     'comments',
     'common',
 ]
