+++ config/urls.py
@@ -22,5 +22,6 @@ urlpatterns = [
     
     path('users/', include("users.urls")),
     path('articles/', include("articles.urls")),
+    path('categories/', include("categories.urls")),
     path('articles/<int:article_id>/comments/', include("comments.urls")),
 ]
